requirements.txt
streamlit
langchain
langchain-cohere
cohere
python-dotenv
sqlalchemy
requests
langsmith
psycopg2-binary 